🗂  세로 가운데정렬 파일 분리
